[SPARKL-45] - Fixed filename not being sent when Content-Disposition was inline

+++ cpk-core/src/pt/webdetails/cpk/utils/ZipUtil.java
@@ -199,7 +199,7 @@ public class ZipUtil {
     return zipName;
   }
 
-  public long getZipSize() {
+  public int getZipSize() {
     try {
       return fis.available();
     } catch ( IOException ex ) {
